lambda/sqs_embedding_task_consumer/pyproject.toml
[project]
name = "sqs-embedding-task-consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "boto3>=1.39.10",
  "psycopg2-binary>=2.9.10",
  "twelvelabs>=1.0.0b0",
]
